Add core base classes and update imports
